pyproject.toml
[project]
name = "mnist-model-collapse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.3.1",
    "matplotlib>=3.10.8",
    "pyyaml>=6.0.1",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
]
